Move System to yash_env::system
The System trait is getting bigger and bigger. Let's tidy up the
namespace by moving the definition to a module.
