+++ yash-env/src/real_system.rs
@@ -16,7 +16,7 @@
 
 //! Implementation of `System` that actually interacts with the system.
 
-use super::ChildProcess;
+use super::system::ChildProcess;
 use super::Env;
 use super::System;
 use crate::io::Fd;
